Delete old main.rs and renamed test_bin_ipc.rs to main.rs
